Changes to support basin outlining
